requirements.txt
Flask
spacy
pdfminer.six
# scrapy
pyresparser
